Adds possibility to connect wallets
The commit adds possibility to connect EVM and Solana wallets.

EVM
I decided to go with `viem`, `wagmi` and `@tanstack/react-query` packages for connecting EVM accounts into the application since they support many different wallets and they are quite easy to be used. Also, to interact with EVM a new wagmi project was created and the id of this project is used now. The id can be found in `.env.example` and I know that this is a bad practise to reveal secret data, but in this case it's fine since it's the home assignment task and the ID should be known by people who are gonna test this application, so I left the id there on purpose. Of course in a real project I wouldn't do this.

Solana
I decoded to go with `@solana/wallet-adapter-react`, `@solana/wallet-adapter-react-ui` and `@solana/wallet-adapter-wallets` for connecting Solana account into the application because those libraries together make interacting with Solana really easy and simple. Also, there is a weird error that pops up when the user connected the account and refreshed the page. I didn't spend enough time on fixing it yet since I think it might take some time and I want to complete all tasks without spending hours on one case. I'll check this error in the end.

Bitcoin
I tried to add the possibility to connect a bitcoin account, but in the end I didn't find any solution for that yet. I used `@bigmi/client`, `@bigmi/core` and `@bigmi/react` libraries, but those libraries don't really have much info in their docs, so it was hard to understand how to work with them and I decided to skip this step so far. I'll come back to this later to try again to don't lose much time right now.

In general, for those features the UI was updated and the structure of the project a bit. The UI is not super fancy, but it reacts on all changes which is enough for now.