Moves repeated UI elements into components
We have components for accounts for 3 different blockchains. All those components use the same UI for some elements, but this same UI wasn't implemented as a component which is a bad practise because it was repeated. So, when almost anything is repeated in programming it must be rewritten.

So, the commit does exactly this. It extracts same UI elements from all 3 account components into base components and those base components are now used across all account components. It's good because styles are now in only one place and if we want to change something for all component we can do it from one place only.