Add tests
The commit adds tests to the application that cover main functionality. Changes that were done for that:
1) Updated jest configuration to match everything in the project;
2) Added mocks to some existing logic for tests to don't rely on external services. Tests even work with no internet;
3) Added a function that renders the given element in tests with all providers that are used in the application. It makes writing tests easier;
4) And finally wrote tests for existing components and pages.

There is only one weird thing, but it might be an issue with my IDE. In some functions from jest types are wrong. For example, `jest.spyOn` expects 3 arguments in my IDE when the jest documentation says that it expects only 2 arguments. And similar cases to that. It doesn't affect tests execution at all.